# [Feature] Allow using NCCL for communication in dgl.NodeEmbedding and dgl.SparseOptimizer