--- a/python/dgl/backend/pytorch/tensor.py
+++ b/python/dgl/backend/pytorch/tensor.py
@@ -337,7 +337,7 @@ def zerocopy_from_dgl_ndarray(data):
         #  Related issue: https://github.com/pytorch/pytorch/issues/41182
         #  The issue will be fixed in v1.6 and later.
         return th.empty(data.shape, dtype=getattr(th, data.dtype),
-                         device=to_backend_ctx(data.ctx))
+                        device=to_backend_ctx(data.ctx))
     else:
         return dlpack.from_dlpack(data.to_dlpack())