--- a/python/dgl/cuda/nccl.py
+++ b/python/dgl/cuda/nccl.py
@@ -88,7 +88,7 @@ class Communicator(object):
                 The 1D set of indices to send to other processors.
             value : tensor
                 The multi-dimension set of values to send to other processors.
-                The 0th dimension must match that of `idx`.
+                The first dimension must match that of `idx`.
             partition : NDArrayPartition
                 The object containing information for assigning indices to
                 processors.