--- a/python/dgl/cuda/nccl.py
+++ b/python/dgl/cuda/nccl.py
@@ -137,7 +137,7 @@ class Communicator(object):
 
     def sparse_all_to_all_pull(self, req_idx, value, partition):
         """ Perform an all-to-all-v operation, where by all processors request
-            the values corresponding to ther set of indices.
+            the values corresponding to their set of indices.
 
             Parameters
             ----------