--- a/python/dgl/nn/pytorch/sparse_emb.py
+++ b/python/dgl/nn/pytorch/sparse_emb.py
@@ -3,8 +3,10 @@ from datetime import timedelta
 import torch as th
 from ...backend import pytorch as F
 from ...utils import get_shared_mem_array, create_shared_mem_array
+from ...cuda import nccl
 
 _STORE = None
+_COMM = None
 
 class NodeEmbedding: # NodeEmbedding
     '''Class for storing node embeddings.
@@ -36,6 +38,11 @@ class NodeEmbedding: # NodeEmbedding
     init_func : callable, optional
         The function to create the initial data. If the init function is not provided,
         the values of the embeddings are initialized to zero.
+    device : th.device
+        Device to store the embeddings on.
+    parittion : String
+        The type of partitioning to use to distributed the embeddings between
+        processes.
 
     Examples
     --------
@@ -58,8 +65,12 @@ class NodeEmbedding: # NodeEmbedding
     '''
 
     def __init__(self, num_embeddings, embedding_dim, name,
-                 init_func=None):
+                 init_func=None, device=None, partition=None):
         global _STORE
+        global _COMM
+
+        if device is None:
+            device = th.device('cpu')
 
         # Check whether it is multi-gpu training or not.
         if th.distributed.is_initialized():
@@ -70,32 +81,81 @@ class NodeEmbedding: # NodeEmbedding
             world_size = 0
         self._rank = rank
         self._world_size = world_size
+        self._store = None
+        self._comm = None
+        self._partition = None
+
         host_name = '127.0.0.1'
         port = 12346
 
-        if rank <= 0:
-            emb = create_shared_mem_array(name, (num_embeddings, embedding_dim), th.float32)
-            if init_func is not None:
-                emb = init_func(emb)
-        if rank == 0: # the master gpu process
+        if rank >= 0:
             # for multi-gpu training, setup a TCPStore for
             # embeding status synchronization across GPU processes
             if _STORE is None:
                 _STORE = th.distributed.TCPStore(
-                    host_name, port, world_size, True, timedelta(seconds=10*60))
-            for _ in range(1, world_size):
-                # send embs
-                _STORE.set(name, name)
-        elif rank > 0:
-            # receive
-            if _STORE is None:
-                _STORE = th.distributed.TCPStore(
-                    host_name, port, world_size, False, timedelta(seconds=10*60))
-            _STORE.wait([name])
-            emb = get_shared_mem_array(name, (num_embeddings, embedding_dim), th.float32)
+                    host_name, port, world_size, rank == 0, timedelta(seconds=10*60))
+            self._store = _STORE
 
-        self._store = _STORE
-        self._tensor = emb
+        if device == th.device('cpu'):
+            if partition:
+                assert self._partition == 'shared'
+            else:
+                partition = 'shared'
+            self._partition = partition
+
+            if rank <= 0:
+                emb = create_shared_mem_array(name, (num_embeddings, embedding_dim), th.float32)
+                if init_func is not None:
+                    emb = init_func(emb)
+            if rank == 0: # the master gpu process
+                for _ in range(1, world_size):
+                    # send embs
+                    self._store.set(name, name)
+            elif rank > 0:
+                # receive
+                self._store.wait([name])
+                emb = get_shared_mem_array(name, (num_embeddings, embedding_dim), th.float32)
+            self._tensor = emb
+        else:
+            if partition:
+                assert self._partition == 'remainder', \
+                        "Only 'remainder' partition scheme is currently supported."
+            else:
+                partition = 'remainder'
+            self._partition = partition
+
+            # setup nccl communicator
+            if _COMM is None:
+                if rank < 0:
+                    _COMM = nccl.Communicator(1, 0, nccl.UniqueId())
+                else:
+                    if rank == 0:
+                        # root process broadcasts nccl id
+                        nccl_id = nccl.UniqueId()
+                        self._store.set('nccl_root_id', str(nccl_id))
+                    else:
+                        nccl_id = nccl.UniqueId(self._store.get('nccl_root_id'))
+                    # needs to be set for nccl to work
+                    th.cuda.set_device(device)
+                    _COMM = nccl.Communicator(self._world_size, self._rank,
+                                              nccl_id)
+            self._comm = _COMM
+
+            # create local tensors for the weights
+            local_size = num_embeddings
+            if rank >= 0:
+                assert self._rank == self._comm.rank()
+                assert self._world_size == self._comm.size()
+
+                local_size = (num_embeddings // world_size) + \
+                    (rank < (num_embeddings % world_size))
+            # TODO(dlasalle): support 16-bit/half embeddings
+            emb = th.empty([local_size, embedding_dim], dtype=th.float32,
+                           requires_grad=False, device=device)
+            if init_func:
+                emb = init_func(emb)
+            self._tensor = emb
+
         self._num_embeddings = num_embeddings
         self._embedding_dim = embedding_dim
         self._name = name
@@ -109,10 +169,19 @@ class NodeEmbedding: # NodeEmbedding
         device : th.device
             Target device to put the collected embeddings.
         """
-        emb = self._tensor[node_ids].to(device)
+        if not self._comm or self._comm.size() == 1:
+            emb = self._tensor[node_ids].to(device)
+        else:
+            if self.world_size > 0:
+                emb = self._comm.sparse_all_to_all_pull(
+                    node_ids, self._tensor, self._partition)
+            else:
+                emb = self._tensor[node_ids]
+            emb = emb.to(device)
         if F.is_recording():
             emb = F.attach_grad(emb)
             self._trace.append((node_ids.to(device, non_blocking=True), emb))
+
         return emb
 
     @property
@@ -126,6 +195,31 @@ class NodeEmbedding: # NodeEmbedding
             KVStore used for meta data sharing.
         """
         return self._store
+
+    @property
+    def comm(self):
+        """Return dgl.cuda.nccl.Communicator for data
+        sharing across processes.
+
+        Returns
+        -------
+        dgl.cuda.nccl.Communicator
+            Communicator used for data sharing.
+        """
+        return self._comm
+
+    @property
+    def partition_mode(self):
+        """Return String for identifying how the tensor is split across
+        processes.
+
+        Returns
+        -------
+        String
+            The mode.
+        """
+
+        return self._partition
 
     @property
     def rank(self):