--- a/python/dgl/nn/pytorch/sparse_emb.py
+++ b/python/dgl/nn/pytorch/sparse_emb.py
@@ -97,6 +97,7 @@ class NodeEmbedding: # NodeEmbedding
                     host_name, port, world_size, rank == 0, timedelta(seconds=10*60))
             self._store = _STORE
 
+        # embeddings is stored in CPU memory.
         if th.device(device) == th.device('cpu'):
             if rank <= 0:
                 emb = create_shared_mem_array(name, (num_embeddings, embedding_dim), th.float32)