… examples/pytorch/rgcn/entity_classify_mp.py
… examples/pytorch/rgcn/model.py
… python/dgl/partition.py
… src/partition/cuda/partition_op.cu
… src/partition/ndarray_partition.cc
… src/partition/ndarray_partition.h
… src/partition/partition_op.h
… src/runtime/cuda/nccl_api.cu
… tests/cpp/test_partition.cc
… tests/pytorch/test_optim.py