## Description
This PR implements keeping embedding weights split across GPUs, and using NCCL to communicate those embeddings during forward, and communicate the gradients during backwards.

There are two changes that are needed in the future that are not part of this PR:
- Allowing more flexible partitioning. That is, rather than only using remainder to assign indexes to processors, allow a range based partition, where a vector part_ptr[N_PROCS+1] could be passed in, and each index would be assigned to a processor based on that.
- Allowing using NCCL for communication in the SparseOptimizer, while storing the embedding the CPU in shared memory.

This PR does not include the actual NCCL implementation, as that has been split into another PR.

## Checklist
Please feel free to remove inapplicable items for your PR.
- [x] The PR title starts with [$CATEGORY] (such as [NN], [Model], [Doc], [Feature]])
- [x] Changes are complete (i.e. I finished coding on this PR)
- [ ] All changes have test coverage
- [x] Code is well-documented
- [x] To the my best knowledge, examples are either not affected by this change,
      or have been fixed to be compatible with this change

## Changes
This adds an extra device parameter to the NodeEmbedding, which allows storing it the on the GPU and using the NCCL wrapper.

This modifies the SparseOptimizer, to check of the embedding has a nccl communicator, and if it does, utilizes that for communicating the gradients.

This modifies the /examples/pytorch/rgcn/entity_classify_mp.py example to accept the --embedding-gpu flag to make use of the new code.

I did some preliminary performance testing with this PR (and the NCCL wrapper implementation).
RGCN fanout=4,4, n_hidden=256, batch_size=1024, ogbn-mag

| GPUS | Shared Memory (forward) | Shared Memory (backward) | NCCL (forward) | NCCL (backward) |
| ---------|----------|-----------|---------|---------|
| 2x V100 | 6.5147s | 13.8149s | 4.4023s | 3.4352s |
| 4x V100 | 6.7430s | 37.9233s | 2.3865s | 1.9602s |

This was done with adding barriers before the forward and backward sections, so as to avoid capturing load imbalances from other steps. One thing to note, is a big part of the speedup comes from storing data on the GPU, and avoid passing data over PCIe and through the CPU all together, where as the shared memory implementation must deal with lower PCIe bandwidth, as well as CPU resource contention.