@classicsong 2021-04-15
> @classicsong I haven't implemented that part yet. I see two options:
> 
> 1. We could add a flag to the sparse optimizer's constructor, telling it exchange gradients on the GPU despite the embedding being the CPU. It would then proceed to create it's own NCCL communicator.
> 2. In the first step of the optimizer, it could see if the gradients it received are on the GPU, create a NCCL communicator there, and the proceed to exchange the gradients.
>    Let me know which you would prefer, and whether you want the change in this PR, or a follow up PR. Thanks.

I prefer the second. The gradients exchange can be decoupled with gradients update. If the gradients are in GPU at the very beginning, we use NCCL communicator to exchange gradients.