@nv-dlasalle 2021-05-24
> I prefer the second. The gradients exchange can be decoupled with gradients update. If the gradients are in GPU at the very beginning, we use NCCL communicator to exchange gradients.

I've implemented the second option now in sparse_optim.py.