@nv-dlasalle 2021-05-26
> 
> 
> Can you add a save_tensor or get_tensor API for NodeEmbedding?
> This is useful when someone wants to save the initial embedding while the embedding is actually store in multiple GPUs.
Do you have any thoughts on the strategy for this? Should we allocate a shared memory tensor, and then have all GPUs copy their sections to that tensor? How would this work for DistDGL where the tensor can't find into a single machines memory?