@classicsong 2021-06-02
> > Can you add a save_tensor or get_tensor API for NodeEmbedding?
> > This is useful when someone wants to save the initial embedding while the embedding is actually store in multiple GPUs.
> 
> @classicsong Do you have any thoughts on the strategy for this? Should we allocate a shared memory tensor, and then have all GPUs copy their sections to that tensor? How is this currently done in DistDGL?

We can allocate a shared memory tensor to have all GPUs copy their sections to that tensor.
For DistDGL, we need some extra discussion about it.